% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{check.spectrum}
\alias{check.spectrum}
\title{An internal function to transit spectrum into 96 channel}
\usage{
check.spectrum(spectrum)
}
\arguments{
\item{spectrum}{A list of possibilities for each mutation type(mutational spectrum) with mutation types as rownames}
}
\value{
A matrix with contribution to the gain-binding and loss-binding for each of 96 mutation types
}
\description{
An internal function to transit spectrum into 96 channel
}
\keyword{internal}
