% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Generate96ChannelSpectrumFromTwelvemers}
\alias{Generate96ChannelSpectrumFromTwelvemers}
\title{Generate mutation spectrum from twelvemers}
\usage{
Generate96ChannelSpectrumFromTwelvemers(twelvemers)
}
\arguments{
\item{twelvemers}{A list of twelvemers, with a 11mer centered at the mutation base and 1 mutated base appended}
}
\value{
A mutation spectrum with 96 mutation types
}
\description{
This function generates a mutation spectrum of 96 mutation classes (mutations on pyrimidine centered trinucleotide) from a list of twelvemers
}
