% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Generate_96Channel_Spectrum_From_Sample_Twelvemers}
\alias{Generate_96Channel_Spectrum_From_Sample_Twelvemers}
\title{Generate mutation spectrum from twelvemers}
\usage{
Generate_96Channel_Spectrum_From_Sample_Twelvemers(twelvemers.list)
}
\arguments{
\item{twelvemers.list}{A list of twelvemers, with a 11mer centered at the mutation base and 1 mutated base appended}
}
\value{
A mutation spectrum with 96 mutation types
}
\description{
Generate mutation spectrum from twelvemers
}
