% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GenerateNormalizedScoresFromTwelvemers}
\alias{GenerateNormalizedScoresFromTwelvemers}
\title{Generate QBiC scores for all uPBMs for a given list of twelvemers}
\usage{
GenerateNormalizedScoresFromTwelvemers(twelvemers.list, PBMscores)
}
\arguments{
\item{twelvemers.list}{A list of twelvemers, with a 11mer centered at the mutation base and 1 mutated base appended}

\item{output.file}{The filename of scores}
}
\value{
Writing the scores into a specific directory
}
\description{
This function generates QBiC scores for all uPBM experiments for given twelvemers
}
