% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GenerateQBiCScoresFromTwelvemers}
\alias{GenerateQBiCScoresFromTwelvemers}
\title{Generate QBiC scores for all uPBMs for a given list of twelvemers}
\usage{
GenerateQBiCScoresFromTwelvemers(uPBM_QBiC_scores, twelvemers.list)
}
\arguments{
\item{uPBM_QBiC_scores}{All QBiC scores for a universal PBM}

\item{twelvemers.list}{A list of twelvemers, with a 11mer centered at the mutation base and 1 mutated base appended}
}
\value{
A list of scores for the given twelvemers list
}
\description{
This function generates QBiC scores for all uPBM experiments for given twelvemers
}
