% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GenerateSignatureWeightedQBiCScoresDistribution}
\alias{GenerateSignatureWeightedQBiCScoresDistribution}
\title{Plot histogram for each mutation class for a given signature and uPBM experiment name}
\usage{
GenerateSignatureWeightedQBiCScoresDistribution(uPBM_QBiC_scores, spectrum,
  output_name)
}
\arguments{
\item{uPBM_QBiC_scores}{All QBiC scores for a PBM}

\item{spectrum}{A list of GR/LR generated by Signature-QBiC from observed mutation spectrum}
}
\value{
A plot of histograms
}
\description{
Plot histogram for each mutation class for a given signature and uPBM experiment name
}
