% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Mutation_Type_Contribution}
\alias{Mutation_Type_Contribution}
\title{Scatter plot with linear regression between prediction and observation}
\usage{
Mutation_Type_Contribution(uPBM_QBiC_scores, signature)
}
\arguments{
\item{signature}{A character of signature}

\item{PBMname}{A character of name of PBM experiment}
}
\value{
A scatter plot with linear regression line
}
\description{
Scatter plot with linear regression between prediction and observation
}
