% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MutationTypeContribution}
\alias{MutationTypeContribution}
\title{Compute the contribution}
\usage{
MutationTypeContribution(uPBM_QBiC_scores, spectrum)
}
\arguments{
\item{uPBM_QBiC_scores}{QBiC scores of a uPBM experiment}

\item{spectrum}{A list of possibilities for each mutation type(mutational spectrum) or a character of name of mutational signature(mutational signature)}
}
\value{
A matrix with contribution to the gain-binding and loss-binding for each of 96 mutation types
}
\description{
Compute the contribution
}
