% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Pathway_Selection}
\alias{Pathway_Selection}
\title{Identify pathways that enriched with affected TFs by a given mutational signature}
\usage{
Pathway_Selection(TF.list, qvalue, dbs.selected)
}
\arguments{
\item{TF.list}{A list of affected TFs}

\item{qvalue}{A number of qvalue as filter}

\item{dbs.selected}{A character of database name in 'enrichR'. In our paper, we used 'Reactome_2016'}
}
\value{
TODO
}
\description{
A function gives the pathways enriched with affected TFs by a given mutational signature
}
