% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PathwaySelection}
\alias{PathwaySelection}
\title{Identify pathways that enriched with affected TFs by a given mutational signature}
\usage{
PathwaySelection(TF.list, qvalue, dbs.selected)
}
\arguments{
\item{TF.list}{A list of affected TFs}

\item{qvalue}{A numeric of qvalue as filter}

\item{dbs.selected}{A character of database name in 'enrichR'. In our paper, we used 'Reactome_2016'}
}
\value{
A matrix of pathways enriched with given TFs
}
\description{
A function gives the pathways enriched with affected TFs by a given mutational signature
}
