% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PlotComparisonPlot}
\alias{PlotComparisonPlot}
\title{Scatter plot with linear regression between prediction and observation}
\usage{
PlotComparisonPlot(Prediction, Observation)
}
\arguments{
\item{Prediction}{A list of GR/LR generated by Signature-QBiC from mutational signatures}

\item{Observation}{A list of GR/LR generated by Signature-QBiC from observed mutation spectrum}
}
\value{
A scatter plot with linear regression line
}
\description{
This function generates scatter plot for GRs and LRs of predictions and observations
}
