% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PyrPenta}
\alias{PyrPenta}
\title{Convert mutations to pyrimidine centred.}
\usage{
PyrPenta(mutstring)
}
\arguments{
\item{mutstring}{A string with for characters: proceeding base - reference base - following base - mutated base (For example, CGA > CCA was input as CGAC)}
}
\value{
A pyrimidine centred mutation
}
\description{
This function converts trinucleotide context mutations to pyrimidine centred. For example, CGA > CAA will be converted to TCG > TTG
}
