% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleMutationFrequency.R
\name{ResampleMutationFrequency}
\alias{ResampleMutationFrequency}
\title{The resampling test generate 1000 flat signatures, to test if the GR or LR>1 is statistically significant.}
\usage{
ResampleMutationFrequency(i)
}
\arguments{
\item{i}{Seednumber}
}
\value{
A list of frequencies of each mutation type with sum to 1
}
\description{
This function tests whether $D'_{Pos}$ ($D'_{Neg}$) are statistically > than $D_{Pos}$ ($D_{Neg}$)
}
