% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Signature_QBiC_GR_LR_Generation}
\alias{Signature_QBiC_GR_LR_Generation}
\title{Generate GR and LR for a uPBM and a signature (Signature-QBiC)}
\usage{
Signature_QBiC_GR_LR_Generation(uPBM_QBiC_scores, signature, type)
}
\arguments{
\item{signature}{A list of possibilitie for each mutation type(mutational spectrum) or a character of name of mutational signature(mutational signature)}

\item{type}{'mutational spectrum' or 'mutational signature'. Requires a rowname with mutation types for mutational spectrum}

\item{PBM.scores}{All QBiC scores for a PBM}
}
\value{
A list of GR and LR
}
\description{
Generate GR and LR for a uPBM and a signature (Signature-QBiC)
}
