% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Signature_QBiC_Plot_Generation}
\alias{Signature_QBiC_Plot_Generation}
\title{Plot histogram for each mutation class for a given signature and uPBM experiment name}
\usage{
Signature_QBiC_Plot_Generation(uPBM_QBiC_scores, signature, output_name)
}
\arguments{
\item{signature}{A list of GR/LR generated by Signature-QBiC from observed mutation spectrum}

\item{PBMname}{TODO}
}
\value{
TODO
}
\description{
Plot histogram for each mutation class for a given signature and uPBM experiment name
}
