% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SignatureQBiC}
\alias{SignatureQBiC}
\title{Generate GR and LR for a uPBM and a signature (Signature-QBiC)}
\usage{
SignatureQBiC(uPBM_QBiC_scores, spectrum)
}
\arguments{
\item{uPBM_QBiC_scores}{All QBiC scores for a universal PBM}

\item{spectrum}{A list of possibilities for each mutation type(mutational spectrum) or a character of name of mutational signature(mutational signature)}
}
\value{
A list of GR and LR
}
\description{
SignatureQBiC model. This function generates Gain Ratio and Loss Ratio for a TF (represented by universal PBM) with a mutational signature (or a mutation spectrum)
}
