% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncatedHist.R
\name{TruncatedHist}
\alias{TruncatedHist}
\title{Plot histograms with the long tails collected into a single bar}
\usage{
TruncatedHist(all.QBiC.scores, original.scores, weighted.prop, mutation.type)
}
\arguments{
\item{all.QBiC.scores}{A numeric vector containing all QBiC scores for a universal PBM.}

\item{original.scores}{A numeric vector containing all QBiC scores for the target mutation type}

\item{weighted.prop}{A numeric variable. The probability of the target mutation type in the mutational signature}

\item{mutation.type}{A character. E.g. "ACA_ATA"}
}
\value{
A list containing GR and LR.
}
\description{
This is an internal function called by SignatureQBiC. It takes
}
