% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTypeContribution.R
\name{MutationTypeContribution}
\alias{MutationTypeContribution}
\title{Compute the contribution}
\usage{
MutationTypeContribution(uPBM_QBiC_scores, p_values, spectrum)
}
\arguments{
\item{uPBM_QBiC_scores}{All QBiC scores for a universal PBM}

\item{p_values}{p_values for all twelvemers}

\item{spectrum}{A list of possibilities for each mutation type(mutational spectrum) or a character of name of mutational signature(mutational signature)}
}
\value{
A matrix with contribution to the GR and LR for each of 96 mutation types
}
\description{
Compute the contribution
}
